pattern-build/tasks/template_runner_local/requirements.txt
requests
pyhcl
retrying
